refactor : 폴더명에 s 제거
