+++ component/note/NoteListCard.tsx
